Reapply "[OpenMP][IRBuilder] Perform finalization (incl. outlining) late"
Reapply e857f098f40dae202361bff9225c43a71e553733 with minor fixes.

The problem was that cancellation can cause new edges to the parallel
region exit block which is not outlined. The CodeExtractor will encode
the information which "exit" was taken as a return value. The fix is to
ensure we do not return any value from the outlined function, to prevent
control to value conversion we ensure a single exit block for the
outlined region.

This reverts commit 55dc9294e48bde8f675fe286fd2810235bca75b0.